Add files via upload
I did not write or claim to have written the HTML or CSS for this project, it was provided to me by my coding bootcamp instructor. I only wrote part of the javascript.

For this project we were given almost all of the code and we just had to pick a free video and play a preview of the video when the mouse hovers over the video and make the preview stop when we move the mouse off of the video.